Add an AbortSignal to IdleOptions (#23)
An `AbortSignal` allows the `IdleDetector` and any other operations to be stopped with a single call to `AbortController.abort()`. So that a single detector can be restarted the options have been moved from the constructor to the `start()` method.

`IdleOptions` to update the internal state of the `IdleDetector` with an optional new threshold or `AbortSignal`.